dinput: return fake DIPROP_GUIDANDPATH property for linux joysticks
Some Ubisoft titles use the property to check if a dinput gamepad is
also an xinput device.

Signed-off-by: Alexey Prokhin <alexey@prokhin.ru>
---
A similar patch for the SDL backend has been already accepted into
Proton. And I have been asked to adapt it for other backends.

is_xinput_device() is based on the code from winebus.sys. In future,
the xinput device detection should be kept in sync as much as possible.