xinput: Fix vibration capabilities
wLeftMotorSpeed and wRightMotorSpeed returned by XInputGetCapabilities
are supposed to be resolutions rather than current motor speeds.

This fixes rumble in Far Cry 5.

Signed-off-by: Alexey Prokhin <alexey@prokhin.ru>
---
I have two concernes about the patch:
1. I have no idea how to write a test for it. The only thing I can
think of is to dump Vibration the same way Gamepad is dumped in
test_get_capabilities(). I can do it in the next revision, if you wish.
2. The patch is a bit incorrect for xinput9_1_0. XInputGetCapabilities
from the library returns wLeftMotorSpeed=wRightMotorSpeed=1. It is
possible to define XINPUT_VER in Makefile.in. But should I bother with
this? Far Cry 5, my testing target, does not seem to care if the
returned values are 1 or 255. The rumble starts working either way.